Fix the broken Spring BeanBuilder syntax to correctly specify the route builder beans for the camel context.
This ended up being the reason why the upgrade to Camel 2.0 was not working. Pre 2.0, Camel would scan _all_ packages looking for RouteBuilders if the context definition did not specify any. The result of the incorrect syntax was that none were being defined, but the automagic was adding them anyway via the scanning mechanism and masking the problem. This magic behaviour was removed from Camel in 2.0.