Configure the camelContext to NOT start until the Grails app has completely started up.
The default behaviour of binding to the Spring context listener events for context startup don't really work in the Grails environment. The camelContext was starting before Grails was ready which meant domain classes didn't have their dynamic methods and all sorts of weird stuff.

To get around this we configure the camelContext not to start when the contextListener says the app has started. To do this we have to step away from the custom ‘camel’ namespace in the Spring DSL and hand craft the beans. Luckily this is trivial.